requirements.txt
# Data handling and preprocessing
pandas==2.1.4
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0

# Image processing
opencv-python==4.8.1.78

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Progress bars
tqdm==4.66.1

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# AWS S3 / Minio
boto3==1.34.10

# Logging
loguru==0.7.2

# Deep Learning - PyTorch
torch==2.1.1
torchvision==0.16.1
torchmetrics==1.2.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Model utilities
scikit-learn==1.3.2

# MLflow Tracking
mlflow==2.9.2
psutil==5.9.6
